% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp_century.R
\name{decomp_century}
\alias{decomp_century}
\title{Decomposition - CENTURY}
\usage{
decomp_century(forc_st, psi, DOC, ACTIVE, SLOW, PASSIVE, timestep)
}
\arguments{
\item{forc_st}{Soil temperature (K)}

\item{psi}{Soil water potential, MPa}

\item{DOC}{Dissolved organic carbon pool}

\item{ACTIVE}{Active soil organic carbon pool}

\item{SLOW}{Slow soil organic carbon pool}

\item{PASSIVE}{Passive soil organic carbon pool}

\item{timestep}{Timestep (days)}
}
\value{
A list containing DOC, ACTIVE,
SLOW, PASSIVE, f_ACTIVE_ATM, f_PASSIVE_ATM,
f_SLOW_ATM, f_ACTIVE_DOC, f_ACTIVE_SLOW, f_SLOW_PASSIVE,
f_ACTIVE_PASSIVE, f_PASSIVE_ACTIVE, f_DOC_Leaching, and f_DOC_ATM.
}
\description{
Decomposition - CENTURY
}
\examples{
decomp_century(forc_st = 15, psi = -0.5, DOC = 0.1, ACTIVE = 1, SLOW = 10, PASSIVE = 100, 1)
}
