% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{Millenial model main loop}
\usage{
main(nr)
}
\arguments{
\item{nr}{Number of simulation steps (days)}
}
\value{
Simulation outputs in a data frame.
}
\description{
Millenial model main loop
}
\examples{
main(365)  # a one-year simulation
}
