% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilpsi.R
\name{soilpsi}
\alias{soilpsi}
\title{Soil hydrological properties}
\usage{
soilpsi(sand, clay, silt, vwc, vwcsat, organic)
}
\arguments{
\item{sand}{Sand (percent)}

\item{clay}{Clay (percent)}

\item{silt}{Silt (percent)}

\item{vwc}{Volumetric water content (m3/m3)}

\item{vwcsat}{Volumetric water content at saturation (m3/m3)}

\item{organic}{Organic matter content (kg/m3)}
}
\value{
A list containing psisat (soil water potential at saturation, MPa);
 psi (soil water potential, MPa); and smp_l (soil matrix potential; mm).
}
\description{
Soil hydrological properties
}
