% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomp_century.R
\name{decomp_century}
\alias{decomp_century}
\title{Decomposition - CENTURY}
\usage{
decomp_century(forc_st, psi, forc_exoenzyme, clay, DOC, ACTIVE, SLOW,
  PASSIVE, f_DOC_ATM, timestep)
}
\arguments{
\item{forc_st}{Soil temperature (K)}

\item{psi}{Soil water potential at saturation, MPa}

\item{forc_exoenzyme}{forc_exoenzyme}

\item{clay}{clay}

\item{DOC}{DOC}

\item{ACTIVE}{ACTIVE}

\item{SLOW}{SLOW}

\item{PASSIVE}{PASSIVE}

\item{f_DOC_ATM}{f_DOC_ATM}

\item{timestep}{Timestep (days)}
}
\value{
A list containing DOC, ACTIVE,
SLOW, PASSIVE, f_ACTIVE_ATM, f_PASSIVE_ATM,
f_SLOW_ATM, f_ACTIVE_DOC, f_ACTIVE_SLOW, f_SLOW_PASSIVE,
f_ACTIVE_PASSIVE, f_PASSIVE_ACTIVE, and f_DOC_Leaching.
}
\description{
Decomposition - CENTURY
}
