% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilpsi.R
\name{soilpsi}
\alias{soilpsi}
\title{Compute soil hydrological properties.}
\usage{
soilpsi(sand, clay, silt, vwc, vwcsat, organic, quiet = FALSE)
}
\arguments{
\item{sand}{Sand (percent)}

\item{clay}{Clay (percent)}

\item{silt}{Silt (percent)}

\item{vwc}{Volumetric water content (m3/m3)}

\item{vwcsat}{Volumetric water content at saturation (m3/m3)}

\item{organic}{Organic matter content (kg/m3)}

\item{quiet}{Print diagnostics?}
}
\value{
A list containing \code{psisat} (soil water potential at saturation, MPa);
 \code{psi} (soil water potential, MPa); and \code{smp_l} (soil matrix potential; mm).
}
\description{
Compute soil hydrological properties.
}
\details{
The calculations generally follow Clapp & Hornberger (1978),
"Empirical equations for some soil hydraulic properties",
\url{https://doi.org/10.1029/WR014i004p00601}. Peatland calculations are based on
Letts et al. (2000), "Parametrization of peatland hydraulic properties
for the Canadian land surface scheme",
\url{https://doi.org/10.1080/07055900.2000.9649643}.
}
\examples{
soilpsi(sand = 40, silt = 40, clay = 20, vwc = 0.4, vwcsat = 0.8, organic = 10)
}
