% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soilpsi.R
\name{soilpsi}
\alias{soilpsi}
\title{Compute soil hydrological properties.}
\usage{
soilpsi(sand, clay, silt, vwc, vwcsat, organic)
}
\arguments{
\item{sand}{Sand (fraction)}

\item{clay}{Clay (fraction)}

\item{silt}{Silt (fraction)}

\item{vwc}{Volumetric water content (m3/m3)}

\item{vwcsat}{Volumetric water content at saturation (m3/m3)}

\item{organic}{Organic matter content (kg/m3)}
}
\value{
A list containing \code{psisat} (soil water potential at saturation, MPa);
 \code{psi} (soil water potential, MPa); and \code{smp_l} (soil matrix potential; mm).
}
\description{
Compute soil hydrological properties.
}
